Extracted template lookup and location into a Locator class.
