+++ src/Titon/View/Locator/TemplateLocator.hh
@@ -1,0 +1,12 @@
+<?hh // strict
+/**
+ * @copyright   2010-2015, The Titon Project
+ * @license     http://opensource.org/licenses/bsd-license.php
+ * @link        http://titon.io
+ */
+
+namespace Titon\View\Locator;
+
+class TemplateLocator extends AbstractLocator {
+
+}
